Add a warning when using `ipytest.autoconfig(coverage=True)` and existing configs
