+++ ipytest/cov.py
@@ -40,8 +40,6 @@ import linecache
 import os
 import os.path
 import re
-
-from pathlib import Path
 from typing import Optional
 
 import coverage.parser
@@ -52,34 +50,6 @@ __all__ = ["translate_cell_filenames"]
 
 _cell_filenames_tracker = None
 config_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "coveragerc")
-
-
-def find_coverage_configs(root):
-    root = Path(root)
-
-    result = []
-    if (p := root.joinpath(".coveragerc")).exists():
-        result.append(p)
-
-    result += _find_files_with_lines(root, ["setup.cfg", "tox.ini"], r"^\[coverage:.*$")
-    result += _find_files_with_lines(root, ["pyproject.toml"], r"^\[tool\.coverage.*$")
-
-    return result
-
-
-def _find_files_with_lines(root, paths, pat):
-    for path in paths:
-        path = root.joinpath(path)
-        if path.exists():
-            try:
-                with open(path, "rt") as fobj:
-                    for line in fobj:
-                        if re.match(pat, line) is not None:
-                            yield path
-                            break
-
-            except Exception:
-                pass
 
 
 def translate_cell_filenames(enabled=True):
